move struct to baseclass
